Impedance type change, name change, bug fix (#1019)
* change impedance type to float

* rename impedance_transmit: impedance_transducer, impedance_receive: impedance_transceiver

* rename: impedance_transmit to impedance_transducer, impedance_receive to impedance_transceiver, Anderson to Anderson, Macaulay to Matlab

* fix line length

* add type hint for z_et and z_er

* document source of gain correction based on angle offset and beamwidth

* remove the "Matlab" implementation of chirp replica

+++ echopype/convert/utils/ek_raw_parsers.py
@@ -851,7 +851,7 @@ class SimradXMLParser(_SimradDatagramParser):
                                     ),
                                     "gain": np.array([float(f.attrib["Gain"]) for f in f_par]),
                                     "impedance": np.array(
-                                        [int(f.attrib["Impedance"]) for f in f_par]
+                                        [float(f.attrib["Impedance"]) for f in f_par]
                                     ),
                                     "phase": np.array([float(f.attrib["Phase"]) for f in f_par]),
                                     "beamwidth_alongship": np.array(
